Refactoring the maze file
