+++ app/maze/page.tsx
@@ -2,10 +2,11 @@
 
 import { ReturnButton } from "@/components/ReturnButton";
 import { useSearchParams } from "next/navigation";
-import { Maze } from "../maze";
+import Maze from "./maze"; // 
+
 
 export default function Mazepage() {
-  const searchParams = useSearchParams();
+  const searchParams = useSearchParams(); 
   const difficulty = searchParams.get("difficulty");
 
   console.log(difficulty);
